+++ scripts/make_summary.py
@@ -318,7 +318,7 @@ def calculate_supply(n, label, supply):
         for c in n.iterate_components(n.one_port_components):
             items = c.df.index[c.df.bus.map(bus_map).fillna(False)]
 
-            if len(items) == 0:
+            if len(items) == 0 or len(c.pnl.p.columns) == 0:
                 continue
 
             s = (
@@ -369,7 +369,7 @@ def calculate_supply_energy(n, label, supply_energy):
         for c in n.iterate_components(n.one_port_components):
             items = c.df.index[c.df.bus.map(bus_map).fillna(False)]
 
-            if len(items) == 0:
+            if len(items) == 0 or len(c.pnl.p.columns) == 0:
                 continue
 
             s = (
@@ -424,7 +424,7 @@ def calculate_nodal_supply_energy(n, label, nodal_supply_energy):
         for c in n.iterate_components(n.one_port_components):
             items = c.df.index[c.df.bus.map(bus_map).fillna(False)]
 
-            if len(items) == 0:
+            if len(items) == 0 or len(c.pnl.p.columns) == 0:
                 continue
 
             s = (
@@ -719,7 +719,7 @@ if __name__ == "__main__":
 
     costs_db = prepare_costs(
         snakemake.input.costs,
-        snakemake.params.costs,
+        snakemake.params,
         Nyears,
     )