+++ scripts/make_summary_perfect.py
@@ -288,7 +288,7 @@ def calculate_supply(n, label, supply):
         for c in n.iterate_components(n.one_port_components):
             items = c.df.index[c.df.bus.map(bus_map).fillna(False)]
 
-            if len(items) == 0:
+            if len(items) == 0 or len(c.pnl.p.columns) == 0:
                 continue
 
             s = (
@@ -351,7 +351,7 @@ def calculate_supply_energy(n, label, supply_energy):
         for c in n.iterate_components(n.one_port_components):
             items = c.df.index[c.df.bus.map(bus_map).fillna(False)]
 
-            if len(items) == 0:
+            if len(items) == 0 or len(c.pnl.p.columns) == 0:
                 continue
 
             if c.name == "Generator":
@@ -735,7 +735,7 @@ if __name__ == "__main__":
     nyears = 1
     costs_db = prepare_costs(
         snakemake.input.costs,
-        snakemake.config["costs"],
+        snakemake.params,
         nyears,
     )