+++ scripts/plot_summary.py
@@ -427,7 +427,7 @@ def plot_carbon_budget_distribution(input_eurostat, options):
         co2_cap = (
             supply_energy.loc["co2"].droplevel(0).drop("co2").sum().unstack().T / 1e9
         )
-        co2_cap.rename(index=lambda x: int(x), inplace=True)
+        co2_cap.rename(index=lambda x: int(float(x)), inplace=True)
 
     plt.figure(figsize=(10, 7))
     gs1 = gridspec.GridSpec(1, 1)